+++ kermit/k95/ckoco3.c
@@ -31,6 +31,7 @@
 #include "ckoqnx.h"
 #include "ckoadm.h"
 #include "ckoads.h"
+#include "ckosnd.h"
 #endif /* NOLOCAL */
 
 #include <ctype.h>              /* Character types */
@@ -200,6 +201,7 @@ extern int tn_b_nlm ;           /* TELNET BINARY newline mode */
 extern int tt_crd;              /* Carriage-return display mode */
 extern int tt_lfd;              /* Line-feed display mode */
 extern int tt_bell;             /* How to handle incoming Ctrl-G characters */
+extern int tt_bell_usr;
 #ifdef KUI
 extern int tt_bell_flash;
 extern int user_bell_flash;
@@ -560,7 +562,7 @@ int      dec_upss = TX_8859_1 ;
 int      dec_lang = VTL_NORTH_AM;       /* DEC VT Language = North American */
 int      dec_nrc  = TX_ASCII;           /* DEC NRC for use during NRC Mode  */
 int      dec_kbd  = TX_8859_1;          /* DEC Keyboard character set       */
-
+bool     sound_playing = FALSE;         /* If a sound is playing */
 /*
   Terminal parameters that can also be set externally by SET commands.
   Formerly they were declared and initialized here, and had different
@@ -11043,6 +11045,9 @@ doreset(int x) {                        /* x = 0 (soft), nonzero (hard) */
 		}
 	}
 
+    /* And any playing sounds */
+    sound_playing = FALSE;
+
     tt_type_mode = tt_type ;
 
     decstglt = DECSTGLT_COLOR;
@@ -11207,6 +11212,7 @@ doreset(int x) {                        /* x = 0 (soft), nonzero (hard) */
     tt_autorepeat = TRUE;               /* Keyboard autorepeat ON */
 #endif /* NT */
     tt_wrap = TRUE;                     /* (FALSE for real VT terminal!) */
+    tt_bell = tt_bell_usr;
     send_c1 = send_c1_usr;              /* Don't send C1 controls */
     keylock = FALSE;                    /* Keyboard is not locked */
 
@@ -17295,7 +17301,26 @@ dodcs( void )
                                 );
                             break;
                         } /* 'q' */
+                        case 't': {     /* DECSWBV */
+                            if (ISVT520(tt_type_mode) || ISK95(tt_type_mode) || ISXTERM(tt_type_mode)) {
+                                extern int beepvolume;
+                                int val = 8; /* high */
+                                char buf[10];
 
+                                if (tt_bell == XYB_NONE) {
+                                    val = 1;
+                                }
+#ifdef NT
+                                else if (tt_bell == XYB_AUD | XYB_MIDI) {
+                                    val = beepvolume == BEEP_VOLUME_LOW ? 4 : 8;
+                                }
+#endif
+                                _snprintf(buf, sizeof(buf), "%d t", val);
+                                _snprintf(decrpss, DECRPSS_LEN,
+                                        fmt, 1, buf);
+                                /* off is 1, low is 4, high is 8 */
+                            }
+                        } /* 't' */
                         } /* achar */
                         break;
                     } /* SP */
@@ -27443,6 +27468,50 @@ vtcsi(void)
 
                     break;
                 } /* '}' */
+                case '~':      /* DECPS - Play Sound */
+                    /* Pvolume ; Pduration ; Pnote */
+                    if (ISVT520(tt_type_mode) || ISK95(tt_type_mode)) {
+                        UCHAR velocity;
+                        int duration, note;
+                        int maxnote;
+                        if (k < 1) pn[1] = 0;
+                        if (pn[1] > 7) return;
+                        if (k < 2) pn[2] = 0;
+                        if (pn[2] > 255) return;
+                        if (k < 3) {
+                            k = 3;
+                            pn[k] = 0;
+                        }
+
+                        /* volume (0-7) to MIDI velocity (0-127) */
+                        velocity = pn[1] * 127 / 7;
+
+                        /* duration (0-255) in 1/32nds of a second to
+                         * milliseconds. This won't be perfect, but we don't
+                         * currently have a microsecond sleep function. So it's
+                         * the best we can reasonably do for now.*/
+                        duration = pn[2] * 1000 / 32;
+
+                        /* The VT520 only plays the first note. For K95, we'll
+                         * play more. */
+                        maxnote = ISK95(tt_type_mode) ? k : 3;
+
+                        sound_playing = TRUE;
+                        for (note = 3; note <= maxnote; note++) {
+                            if (!sound_playing) break;
+
+                            /* A bunch of sites give 72 as the number for C5 in
+                             * MIDI, and they progress up from there in the same
+                             * order the VT520 manual gives, so... */
+                            pn[note] += 71;
+                            MakeSound(
+                                pn[note] > 96 ? 0 : pn[note],  /* Note out of range? silent */
+                                pn[note] == 71 ? 0 : velocity, /* note 0 is silent */
+                                duration);
+                        }
+                        sound_playing = FALSE;
+                    };
+                    break;
 				case '|': {    /*  DECAC - Assign Color */
 					if (ISVT525(tt_type_mode)) {
 						cell_video_attr_t att = cell_video_attr_init_vio_attribute(0x00);
@@ -27659,6 +27728,35 @@ vtcsi(void)
                     }
                     setcursormode() ;
                     cursorena[VTERM] = TRUE;
+                    break;
+                case 't':   /* DECSWBV - Set Warning Bell Volume */
+#ifdef NT
+                    if (ISVT520(tt_type_mode) || ISK95(tt_type_mode) || ISXTERM(tt_type_mode)) {
+                        extern int beepvolume;
+                        if (k < 1) pn[1] = 5;
+                        switch ( pn[1] ) {
+                            case 1:
+                                /* off */
+                                tt_bell = XYB_NONE;
+                                break;
+                            case 2:
+                            case 3:
+                            case 4:
+                                /* low */
+                                tt_bell = XYB_AUD | XYB_MIDI;
+                                beepvolume = BEEP_VOLUME_LOW;
+                                break;
+                            case 5:
+                            case 6:
+                            case 7:
+                            case 8:
+                                /* high */
+                                tt_bell = XYB_AUD | XYB_MIDI;
+                                beepvolume = BEEP_VOLUME_HIGH;
+                                break;
+                        }
+                    }
+#endif /* NT */
                     break;
                 }
                 break;