+++ kermit/k95/ckocon.h
@@ -1326,7 +1326,7 @@ typedef struct videoline_struct {
     vt_char_attr_t *    vt_char_attrs ;   /* bitwise & of VT_CHAR_ATTR Values */
 #ifdef KUI
     vt_cell_attr_t *    cell_attrs;       /* bitwise & of CA_ATTR Values */
-#endif /* KUI */    
+#endif /* KUI */
     unsigned short *    hyperlinks;       /* hyperlink index values */
     unsigned short      vt_line_attr ;
     short               markbeg ;
@@ -1923,6 +1923,10 @@ _PROTOTYP( hyperlink * hyperlink_get, (int));
 #else
 #define DEF_BEEP_TIME 100
 #endif
+
+#define DEF_BEEP_VOLUME 127
+#define BEEP_VOLUME_LOW 54
+#define BEEP_VOLUME_HIGH 127
 
 #define ATTR_MODE_REAL 1
 #define ATTR_MODE_INTENSITY 2